# app/internal/storages/postgress/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.28.0

package sqlc

import (
	"time"

	"github.com/google/uuid"
)

type Subscription struct {
	ID               uuid.UUID  `db:"id"`
	ConfirmationCode uuid.UUID  `db:"confirmation_code"`
	Status           string     `db:"status"`
	Frequency        string     `db:"frequency"`
	Email            string     `db:"email"`
	City             string     `db:"city"`
	SendedAt         *time.Time `db:"sended_at"`
}